Fix formatting error
If `LogEventInfo.Parameters` is set, NLog will treat the message string as
a format string and attempt to interpolate it with values from the parameters
array.

This means that if the message string contains a `{1}` or similar parameter
placeholder, and this index falls outside the size of the parameters array,
an `IndexOutOfRangeException` will be thrown.

The logging context was passed as a single parameter for log messages that
did not have an attached exception. This was not necessary (since the contents
of the context are also provided as properties in `LogEventInfo.Properties`)
and so this commit removes it, bringing it in line with the messages that do
have an attached exception.